Was just finishing dual streambuf positions and am thinking it's stupidly complicated
Currently, we look for writes into the buffered get-region and then update the get-region,
but that seems like it's going to be a nest of subtle errors to waste an eternity debugging.
So, maybe I will leave this branch here, merge it, and then simplify this to a version with only
one position pointer, and which demands that all seeks be for both directions, as that is how
UNIX file pointers (and probably other OSes) work. If you don't do that, then you have to
seek every time you switch from reading to writing or vice versa, but then not every file
abstraction is capable of seeking (sockets, etc).

So, then you would need to open a file twice, once for reading and once for writing.
However, I don't think you can open the same socket twice, and dup() does not provide
independent file pointers. Both descriptors share the same pointer. So, then, it's clear
that if we want to be compatible with a range of UNIX files, both seekable and non-seekable,
then we need to assume a single seek-pointer. Having two pointers really does not get you
anything other than complexity, and you can accomplish all of the same things with one.